match semantics of what KHFs do
The `SDBTree` doesn't implement truncation of keys, but that's because
keys in a `SDBTree` may not be derived contiguously. For example, when
we use an `SDBTree` to provide localized keys per inode in our FUSE
passthrough, the key IDs are really `inumber || block`, where `||`
indicates concatenation. It should be apparent that telling the
`SDBTree` to be truncated to a specific number of keys doesn't really
make sense; how do you decide which keys to be truncated?